Move all of the view helpers into a directory inside of `views`
