+++ src/views/helpers/Alignment.ts
